refactor: define query functions in mihama.query
