+++ tests/test_query.py
@@ -1,7 +1,7 @@
 import pytest
 
 from mihama import schemas
-from mihama.query import normalize_query
+from mihama.query import batch_query, normalize_query
 
 
 @pytest.fixture
@@ -30,3 +30,14 @@ def test_normalize_query_with_purl_with_version(
 ):
     normalized = normalize_query(query_with_package_with_purl_with_version)
     assert normalized.version is not None
+
+
+@pytest.fixture
+def empty_batch_query():
+    return schemas.BatchQuery(queries=[])
+
+
+@pytest.mark.asyncio
+async def test_batch_query_with_empty_queries(empty_batch_query: schemas.BatchQuery):
+    res = await batch_query(queries=empty_batch_query)
+    assert len(res.results) == 0
